Decouple ODEDynamics class from TorchDiffEq (#290)
* factored out generic Dynamics class from ODEDynamics and moved torchdiffeq specific implementations out of ODEDynamics class

* update dynamics test imports

* moved non user-facing objects and operations to internals.py

* split out operations, handlers, and internals

* moved around ODE operations and handlers

* lint

* lint again

* separated effectful operations into hidden dispatch

* change kwargs in test to args

* added note

* separated out TorchDiffEq Backend

* tidy up tests

* remove comment

* lint

* lint

* added options to TorchDiffEqBackend

* moved handlers.py into folder

* add files to __init__.py for handlers

* lint

* seperated interruptions from simulator event loop

* added placeholder BackendHandler class

* added BackendHandler and edited tests accordingly

* lint

* lint again

* remove unnecessary type check

* renamed backend to solver throughout

* moved functionality out of __init__.py

* updated and reran notebook

* lingering test without solver kwarg. Probably isn't used anywhere?

* missing import
