+++ chirho/dynamical/ops/ODE/solvers/__init__.py
@@ -1,0 +1,1 @@
+from .torchdiffeq import TorchDiffEq  # noqa: F401
