+++ chirho/dynamical/ops/ODE/solvers/torchdiffeq.py
@@ -1,0 +1,15 @@
+from chirho.dynamical.ops.ODE import ODEBackend
+
+
+class TorchDiffEq(ODEBackend):
+    def __init__(self, rtol=1e-7, atol=1e-9, method=None, options=None):
+        self.rtol = rtol
+        self.atol = atol
+        self.method = method
+        self.options = options
+        self.odeint_kwargs = {
+            "rtol": rtol,
+            "atol": atol,
+            "method": method,
+            "options": options,
+        }
