+++ chirho/dynamical/ops/__init__.py
@@ -1,1 +1,1 @@
-from .dynamical import Dynamics, State, Trajectory, simulate  # noqa: F401
+from .dynamical import Dynamics, Solver, State, Trajectory, simulate  # noqa: F401
