+++ tests/dynamical/test_dynamic_interventions.py
@@ -11,6 +11,7 @@ from chirho.counterfactual.handlers import (
 from chirho.dynamical.handlers import DynamicIntervention, SimulatorEventLoop
 from chirho.dynamical.ops import State, simulate
 from chirho.dynamical.ops.ODE import ODEDynamics
+from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 
 from .dynamical_fixtures import UnifiedFixtureDynamics
@@ -81,7 +82,7 @@ def test_nested_dynamic_intervention_causes_change(
                 var_order=init_state.var_order,
                 max_applications=1,
             ):
-                res = simulate(model, init_state, tspan)
+                res = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     preint_total = init_state.S + init_state.I + init_state.R
 
@@ -136,7 +137,7 @@ def test_dynamic_intervention_causes_change(
             var_order=init_state.var_order,
             max_applications=1,
         ):
-            res = simulate(model, init_state, tspan)
+            res = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     preint_total = init_state.S + init_state.I + init_state.R
 
@@ -191,7 +192,9 @@ def test_split_twinworld_dynamic_intervention(
                 max_applications=1,
             ):
                 with TwinWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with cf:
         for k in cf_trajectory.keys:
@@ -230,7 +233,9 @@ def test_split_multiworld_dynamic_intervention(
                 max_applications=1,
             ):
                 with MultiWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with cf:
         for k in cf_trajectory.keys:
@@ -268,7 +273,9 @@ def test_split_twinworld_dynamic_matches_output(
                 max_applications=1,
             ):
                 with TwinWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with SimulatorEventLoop():
         with DynamicIntervention(
@@ -283,10 +290,10 @@ def test_split_twinworld_dynamic_matches_output(
                 var_order=init_state.var_order,
                 max_applications=1,
             ):
-                expected_cf = simulate(model, init_state, tspan)
+                expected_cf = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     with SimulatorEventLoop():
-        expected_factual = simulate(model, init_state, tspan)
+        expected_factual = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     with cf:
         factual_indices = IndexSet(
@@ -346,7 +353,7 @@ def test_grad_of_dynamic_intervention_event_f_params():
     # noinspection DuplicatedCode
     with SimulatorEventLoop():
         with dynamic_intervention:
-            traj = simulate(model, initial_state=s0, timespan=torch.tensor([0.0, 10.0]))
+            traj = simulate(model, s0, torch.tensor([0.0, 10.0]), solver=TorchDiffEq())
 
     (dxdparam,) = torch.autograd.grad(
         outputs=(traj.x[-1],), inputs=(param,), create_graph=True
