+++ tests/dynamical/test_handler_composition.py
@@ -12,6 +12,7 @@ from chirho.dynamical.handlers import (
     SimulatorEventLoop,
 )
 from chirho.dynamical.ops import State, simulate
+from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 from chirho.observational.handlers.soft_conditioning import AutoSoftConditioning
 from tests.dynamical.dynamical_fixtures import (
     UnifiedFixtureDynamics,
@@ -64,6 +65,7 @@ def counterf_model():
                 UnifiedFixtureDynamicsReparam(beta=0.5, gamma=0.7),
                 init_state,
                 tspan,
+                solver=TorchDiffEq(),
             )
