+++ tests/dynamical/test_noop_interruptions.py
@@ -10,6 +10,7 @@ from chirho.dynamical.handlers import (
     SimulatorEventLoop,
 )
 from chirho.dynamical.ops import State, simulate
+from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 
 from .dynamical_fixtures import UnifiedFixtureDynamics, check_trajectories_match
 
@@ -36,12 +37,14 @@ intervene_states = [
 @pytest.mark.parametrize("init_state", [init_state_values])
 @pytest.mark.parametrize("tspan", [tspan_values])
 def test_noop_point_interruptions(model, init_state, tspan):
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     # Test with standard point interruptions within timespan.
     with SimulatorEventLoop():
         with PointInterruption(time=tspan[-1] / 2.0 + eps):
-            result_pint = simulate(model, init_state, tspan)
+            result_pint = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     assert check_trajectories_match(observational_execution_result, result_pint)
 
@@ -51,7 +54,9 @@ def test_noop_point_interruptions(model, init_state, tspan):
             time=tspan[-1] / 4.0 + eps
         ):  # roughly 1/4 of the way through the timespan
             with PointInterruption(time=(tspan[-1] / 4.0) * 3 + eps):  # roughly 3/4
-                result_double_pint1 = simulate(model, init_state, tspan)
+                result_double_pint1 = simulate(
+                    model, init_state, tspan, solver=TorchDiffEq()
+                )
 
     assert check_trajectories_match(observational_execution_result, result_double_pint1)
 
@@ -59,7 +64,9 @@ def test_noop_point_interruptions(model, init_state, tspan):
     with SimulatorEventLoop():
         with PointInterruption(time=(tspan[-1] / 4.0) * 3 + eps):  # roughly 3/4
             with PointInterruption(time=tspan[-1] / 4.0 + eps):  # roughly 1/3
-                result_double_pint2 = simulate(model, init_state, tspan)
+                result_double_pint2 = simulate(
+                    model, init_state, tspan, solver=TorchDiffEq()
+                )
 
     assert check_trajectories_match(observational_execution_result, result_double_pint2)
 
@@ -78,7 +85,9 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
 
     post_measurement_intervention_time = tspan_values.max() + 1.0
 
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     # Test a single point intervention.
     with pytest.warns(
@@ -88,7 +97,9 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
             with PointIntervention(
                 time=post_measurement_intervention_time, intervention=intervene_state
             ):
-                result_single_pi = simulate(model, init_state, tspan)
+                result_single_pi = simulate(
+                    model, init_state, tspan, solver=TorchDiffEq()
+                )
 
     assert check_trajectories_match(observational_execution_result, result_single_pi)
 
@@ -104,7 +115,9 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
                     time=post_measurement_intervention_time + 1.0,
                     intervention=intervene_state,
                 ):
-                    result_double_pi1 = simulate(model, init_state, tspan)
+                    result_double_pi1 = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     assert check_trajectories_match(observational_execution_result, result_double_pi1)
 
@@ -121,7 +134,9 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
                     time=post_measurement_intervention_time,
                     intervention=intervene_state,
                 ):
-                    result_double_pi2 = simulate(model, init_state, tspan)
+                    result_double_pi2 = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     assert check_trajectories_match(observational_execution_result, result_double_pi2)
 
@@ -130,11 +145,13 @@ def test_noop_point_interventions(model, init_state, tspan, intervene_state):
 @pytest.mark.parametrize("init_state", [init_state_values])
 @pytest.mark.parametrize("tspan", [tspan_values])
 def test_point_interruption_at_start(model, init_state, tspan):
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     with SimulatorEventLoop():
         with PointInterruption(time=1.0):
-            result_pint = simulate(model, init_state, tspan)
+            result_pint = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     assert check_trajectories_match(observational_execution_result, result_pint)
 
@@ -144,7 +161,9 @@ def test_point_interruption_at_start(model, init_state, tspan):
 @pytest.mark.parametrize("tspan", [tspan_values])
 @pytest.mark.parametrize("intervene_state", intervene_states)
 def test_noop_dynamic_interruption(model, init_state, tspan, intervene_state):
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     with SimulatorEventLoop():
         tt = (tspan[-1] - tspan[0]) / 2.0
@@ -153,6 +172,6 @@ def test_noop_dynamic_interruption(model, init_state, tspan, intervene_state):
             var_order=init_state.var_order,
             max_applications=1,
         ):
-            result_dint = simulate(model, init_state, tspan)
+            result_dint = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     assert check_trajectories_match(observational_execution_result, result_dint)
