+++ tests/dynamical/test_static_interventions.py
@@ -9,6 +9,7 @@ from chirho.counterfactual.handlers import (
 )
 from chirho.dynamical.handlers import PointIntervention, SimulatorEventLoop
 from chirho.dynamical.ops import State, simulate
+from chirho.dynamical.ops.ODE.solvers import TorchDiffEq
 from chirho.indexed.ops import IndexSet, gather, indices_of
 from chirho.interventional.ops import intervene
 
@@ -50,7 +51,9 @@ eps = 1e-3
 def test_point_intervention_causes_difference(
     model, init_state, tspan, intervene_state, intervene_time
 ):
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     # Simulate with the intervention and ensure that the result differs from the observational execution.
     with SimulatorEventLoop():
@@ -59,10 +62,12 @@ def test_point_intervention_causes_difference(
                 with pytest.raises(
                     ValueError, match="occurred before the start of the timespan"
                 ):
-                    simulate(model, init_state, tspan)
+                    simulate(model, init_state, tspan, solver=TorchDiffEq())
                 return
             else:
-                result_single_pint = simulate(model, init_state, tspan)
+                result_single_pint = simulate(
+                    model, init_state, tspan, solver=TorchDiffEq()
+                )
 
     assert check_trajectories_match_in_all_but_values(
         observational_execution_result, result_single_pint
@@ -103,7 +108,9 @@ def test_nested_point_interventions_cause_difference(
     intervene_state2,
     intervene_time2,
 ):
-    observational_execution_result = simulate(model, init_state, tspan)
+    observational_execution_result = simulate(
+        model, init_state, tspan, solver=TorchDiffEq()
+    )
 
     # Simulate with the intervention and ensure that the result differs from the observational execution.
     with SimulatorEventLoop():
@@ -113,7 +120,7 @@ def test_nested_point_interventions_cause_difference(
                     with pytest.raises(
                         ValueError, match="occurred before the start of the timespan"
                     ):
-                        simulate(model, init_state, tspan)
+                        simulate(model, init_state, tspan, solver=TorchDiffEq())
                     return
                 # AZ - We've decided to support this case and have interventions apply sequentially in the order
                 #  they are handled.
@@ -122,10 +129,12 @@ def test_nested_point_interventions_cause_difference(
                 #         ValueError,
                 #         match="Two point interruptions cannot occur at the same time.",
                 #     ):
-                #         simulate(model, init_state, tspan)
+                #         simulate(model, init_state, tspan, solver=TorchDiffEq())
                 #     return
                 else:
-                    result_nested_pint = simulate(model, init_state, tspan)
+                    result_nested_pint = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     assert check_trajectories_match_in_all_but_values(
         observational_execution_result, result_nested_pint
@@ -152,7 +161,9 @@ def test_twinworld_point_intervention(
                 time=intervene_time + 0.5, intervention=intervene_state
             ):
                 with TwinWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with cf:
         for k in cf_trajectory.keys:
@@ -174,7 +185,9 @@ def test_multiworld_point_intervention(
                 time=intervene_time + 0.5, intervention=intervene_state
             ):
                 with MultiWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with cf:
         for k in cf_trajectory.keys:
@@ -191,7 +204,7 @@ def test_split_odeint_broadcast(
 ):
     with TwinWorldCounterfactual() as cf:
         cf_init_state = intervene(init_state_values, intervene_state, event_dim=0)
-        trajectory = simulate(model, cf_init_state, tspan)
+        trajectory = simulate(model, cf_init_state, tspan, solver=TorchDiffEq())
 
     with cf:
         for k in trajectory.keys:
@@ -213,17 +226,19 @@ def test_twinworld_matches_output(
                 time=intervene_time + 0.543, intervention=intervene_state
             ):
                 with TwinWorldCounterfactual() as cf:
-                    cf_trajectory = simulate(model, init_state, tspan)
+                    cf_trajectory = simulate(
+                        model, init_state, tspan, solver=TorchDiffEq()
+                    )
 
     with SimulatorEventLoop():
         with PointIntervention(time=intervene_time, intervention=intervene_state):
             with PointIntervention(
                 time=intervene_time + 0.543, intervention=intervene_state
             ):
-                expected_cf = simulate(model, init_state, tspan)
+                expected_cf = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     with SimulatorEventLoop():
-        expected_factual = simulate(model, init_state, tspan)
+        expected_factual = simulate(model, init_state, tspan, solver=TorchDiffEq())
 
     with cf:
         factual_indices = IndexSet(
